==> Python/P2-PCA Eigen.ipynb <==
# This notebook calculates the eigenvalues and other useful values of the PCA results by biodivmapR, and then save them into a .csv file in local storage. 

import os
import pandas as pd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_PRISMA = cwd + "\\PRISMA Raster Raw\\Merged"
cwd_PRISMA_Result = cwd + "\\ResultsPRISMA"

# Get the name of all the rasters in our folder
raster_Names = []
for item in [name.split('.')[0] for name in os.listdir(cwd_PRISMA)]:
    if item not in raster_Names:
        raster_Names.append(item)
raster_Names

df_Eigen_6 = pd.read_csv(cwd_PRISMA_Result + "\\" + raster_Names[0] + "\\PCA R.csv", sep = '\t')
df_Eigen_9 = pd.read_csv(cwd_PRISMA_Result + "\\" + raster_Names[1] + "\\PCA R.csv", sep = '\t')
df_Eigen_9n = pd.read_csv(cwd_PRISMA_Result + "\\" + raster_Names[2] + "\\PCA R.csv", sep = '\t')
df_Eigen_6.head()

df_Eigen_6.columns = ['Standard Deviation']
df_Eigen_9.columns = ['Standard Deviation']
df_Eigen_9n.columns = ['Standard Deviation']

list_PC = []
for i in range(len(df_Eigen_6)):
    temp = "PC" + str(i+1)
    list_PC.append(temp)
list_PC[-1]

df_Eigen_6['PC'] = list_PC
df_Eigen_9['PC'] = list_PC
df_Eigen_9n['PC'] = list_PC
df_Eigen_6

df_Eigen_6['Eigenvalue'] = df_Eigen_6['Standard Deviation']**2
df_Eigen_9['Eigenvalue'] = df_Eigen_9['Standard Deviation']**2
df_Eigen_9n['Eigenvalue'] = df_Eigen_9n['Standard Deviation']**2
df_Eigen_6

df_Eigen_6['Proportion'] = (df_Eigen_6["Eigenvalue"] / df_Eigen_6["Eigenvalue"].sum())
df_Eigen_9['Proportion'] = df_Eigen_9["Eigenvalue"] / df_Eigen_9["Eigenvalue"].sum()
df_Eigen_9n['Proportion'] = df_Eigen_9n["Eigenvalue"] / df_Eigen_9n["Eigenvalue"].sum()
# df_Eigen_9['Eigenvalue Percentage'] = df_Eigen_9.Eigenvalue / df_Eigen_9.Eigenvalue.sum()
df_Eigen_6

df_Eigen_6['Cumulative Proportion'] = df_Eigen_6['Proportion'].cumsum()
df_Eigen_9['Cumulative Proportion'] = df_Eigen_9['Proportion'].cumsum()
df_Eigen_9n['Cumulative Proportion'] = df_Eigen_9n['Proportion'].cumsum()
df_Eigen_6

df_Eigen_6 = df_Eigen_6[['PC','Standard Deviation','Eigenvalue','Proportion','Cumulative Proportion']]
df_Eigen_9 = df_Eigen_9[['PC','Standard Deviation','Eigenvalue','Proportion','Cumulative Proportion']]
df_Eigen_9n = df_Eigen_9n[['PC','Standard Deviation','Eigenvalue','Proportion','Cumulative Proportion']]
df_Eigen_6

df_Eigen_6['Standard Deviation'] = df_Eigen_6['Standard Deviation'].round(6)
df_Eigen_9['Standard Deviation'] = df_Eigen_9['Standard Deviation'].round(6)
df_Eigen_9n['Standard Deviation'] = df_Eigen_9['Standard Deviation'].round(6)
df_Eigen_6['Eigenvalue'] = df_Eigen_6['Eigenvalue'].round(6)
df_Eigen_9['Eigenvalue'] = df_Eigen_9['Eigenvalue'].round(6)
df_Eigen_9n['Eigenvalue'] = df_Eigen_9['Eigenvalue'].round(6)
df_Eigen_6['Proportion'] = df_Eigen_6['Proportion'].round(6)
df_Eigen_9['Proportion'] = df_Eigen_9['Proportion'].round(6)
df_Eigen_9n['Proportion'] = df_Eigen_9['Proportion'].round(6)
df_Eigen_6['Cumulative Proportion'] = df_Eigen_6['Cumulative Proportion'].round(6)
df_Eigen_9['Cumulative Proportion'] = df_Eigen_9['Cumulative Proportion'].round(6)
df_Eigen_9n['Cumulative Proportion'] = df_Eigen_9['Cumulative Proportion'].round(6)
df_Eigen_6

df_Eigen_6.to_csv(cwd_PRISMA_Result + "\\" + raster_Names[0] + "\\" + raster_Names[0] + "_PCA_Eigen.csv", index = False)
df_Eigen_9.to_csv(cwd_PRISMA_Result + "\\" + raster_Names[1] + "\\" + raster_Names[1] + "_PCA_Eigen.csv", index = False)
df_Eigen_9n.to_csv(cwd_PRISMA_Result + "\\" + raster_Names[2] + "\\" + raster_Names[2] + "_PCA_Eigen.csv", index = False)
