==> Python/5-1-FieldData1 to CSV.ipynb <==
# This notebook refines the Excel file of the field data collected by our group, keeping only the necessary information and saving them as an easily accessible .csv file. 

import numpy as np
import pandas as pd
import geopandas as gpd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Field_Data1 = cwd_Field + '\\Field Dataset 1'
cwd_Field_Data1_CSV = cwd_Field_Data1 + "\\CSV"

### Refinement

df_Field1 = pd.read_excel(cwd_Field_Data1 + '\\Sintesi_siti_ParcoTicino_con_specie.xlsx')
df_Field1.head()

# Fill NaN with 0 to simplify the later process
df_Field1 = df_Field1.fillna(0)
df_Field1.head()

# Convert the default float values to int64 values
df_Field1_Columns = df_Field1.columns
for column in df_Field1_Columns:
    if ("Piano Dominato" in column) or ("Piano Dominante" in column):
        if not ("Copertura" in column):
            df_Field1 = df_Field1.astype({column: 'int64'})
df_Field1.head()

# Drop useless columns (We drop the columns of Copertura as well since currently these values don't contribute to our work yet. )
df_Field1_Useful = df_Field1.copy()
for column in df_Field1_Columns:
    if not ((column == 'ESU') or ("Piano Dominato" in column) or ("Piano Dominante" in column) or (column == 'Lat') or (column == 'Long') or (column == 'X_UTM (est)') or (column == 'Y_UTM (nord)')) or ('Copertura' in column):
        df_Field1_Useful = df_Field1_Useful.drop(columns = [column])
df_Field1_Useful

# Get the rows where there are no data of the quantities of the species
df_Field1_Columns_Truncated = df_Field1_Useful.columns
list_Row_toDrop = []
for row in range(0,df_Field1_Useful.shape[0]):
    temp_Check_toDrop = True
    for column in df_Field1_Columns_Truncated:
        if ("Piano Dominato" in column) or ("Piano Dominante" in column):
            if df_Field1_Useful.loc[df_Field1_Useful.iloc[row].name,column] != 0:
                temp_Check_toDrop = False
                break
    if temp_Check_toDrop:
        list_Row_toDrop.append(row)
list_Row_toDrop

# Drop those empty rows! 
df_Field1_Valid = df_Field1_Useful.drop(index = df_Field1_Useful.index[list_Row_toDrop])
df_Field1_Valid

# Combine "Piano Dominante" and "Piano Dominato", merging the same species and keeping the unique species
list1_Spezie = []
for spezie in df_Field1_Valid.columns:
    if " - " in spezie:
        list1_Spezie.append(spezie)
list1_Spezie
list_Spezie_Dominato = []
list_Spezie_Dominante = []
for spezie in list1_Spezie:
    if "Dominato" in spezie:
        list_Spezie_Dominato.append(spezie.split(" - ")[1])
    elif "Dominante" in spezie:
        list_Spezie_Dominante.append(spezie.split(" - ")[1])
# Sum up the counts of the same plants in "Piano Dominante" and "Piano Dominato". 
df_Field1_Merge = df_Field1_Valid.copy()
for spezie_Dominante in list_Spezie_Dominante:
    # Set a check value in case that a plant only exists in "Piano Dominante"
    spezie_unique = True
    for spezie_Dominato in list_Spezie_Dominato: 
        if spezie_Dominante == spezie_Dominato:
            # Add up the counts if the plant exists in both "Piano Dominante" and "Piano Dominato"
            df_Field1_Merge[spezie_Dominante] = df_Field1_Merge[("Piano Dominante - " + spezie_Dominante)] + df_Field1_Merge[("Piano Dominato - " + spezie_Dominato)]
            # Drop the two columns! 
            df_Field1_Merge = df_Field1_Merge.drop(columns = [("Piano Dominante - " + spezie_Dominante),("Piano Dominato - " + spezie_Dominato)])
            spezie_unique = False
            break
    # If the plant only exists in "Piano Dominante", then no need to add any arithmetric operation! 
    if spezie_unique:
        df_Field1_Merge[spezie_Dominante] = df_Field1_Merge[("Piano Dominante - " + spezie_Dominante)]
        df_Field1_Merge = df_Field1_Merge.drop(columns = [("Piano Dominante - " + spezie_Dominante)])
# Now only the plants unique in "Piano Dominato" remain intact in the geodataframe. We simply rename their names! 
for column in df_Field1_Merge.columns: 
    if "Piano Dominato" in column:
        df_Field1_Merge = df_Field1_Merge.rename(columns={column: column.split(" - ")[1]})
# Finally check our columns of the final dataframe! 
df_Field1_Merge

# Convert the latitude and longitude to the correct form
df_Field1_Merge['Lat'] = df_Field1_Merge['Lat'].str.split(r'°|\'', regex = True, expand = True).iloc[:,0].apply(pd.to_numeric) + df_Field1_Merge['Lat'].str.split(r'°|\'', regex = True, expand = True).iloc[:,1].apply(pd.to_numeric).apply(lambda row: row/60)
df_Field1_Merge['Long'] = df_Field1_Merge['Long'].str.split(r'°|\'', regex = True, expand = True).iloc[:,0].apply(pd.to_numeric) + df_Field1_Merge['Long'].str.split(r'°|\'', regex = True, expand = True).iloc[:,1].apply(pd.to_numeric).apply(lambda row: row/60)
df_Field1_Merge

# Now we save only the geographical coordinates of all the valid plots! 
df_Field1_Merge[['ESU','Lat','Long']].rename(columns = {'ESU': 'Plot'}).to_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid LatLong.csv", index = False)
df_Field1_Merge[['ESU','X_UTM (est)','Y_UTM (nord)']].rename(columns = {'ESU': 'Plot', 'X_UTM (est)': 'UTM_X', 'Y_UTM (nord)': 'UTM_Y'}).to_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid UTM.csv", index = False)

### Now in this dataframe the names of all the species are columns. Here we want to transfer them into a row. 

# Convert the species columns into rows, so that the new dataframe will only have 5 columns, where the fourth column will be the name of the species and the fifth column will be the quantity of that species
list1_FOR = []
list1_Lat = []
list1_Long = []
list1_UTM_X = []
list1_UTM_Y = []
list1_species = []
list1_quantity = []
for row in range(0,df_Field1_Merge.shape[0]):
    for column in df_Field1_Merge.columns:
        if not (column == 'ESU') and not (column == 'Lat') and not (column == 'Long') and not (column == 'X_UTM (est)') and not (column == 'Y_UTM (nord)'):
            if df_Field1_Merge.loc[df_Field1_Merge.iloc[row].name,column] != 0:
                list1_FOR.append(df_Field1_Merge.iloc[row,0])
                list1_Lat.append(df_Field1_Merge.iloc[row,1])
                list1_Long.append(df_Field1_Merge.iloc[row,2])
                list1_UTM_X.append(df_Field1_Merge.iloc[row,3])
                list1_UTM_Y.append(df_Field1_Merge.iloc[row,4])
                list1_species.append(column)
                list1_quantity.append(df_Field1_Merge.loc[df_Field1_Merge.iloc[row].name,column])
df_Field1_Converted = pd.DataFrame(
    data = {
        'Plot': list1_FOR,
        'Lat': list1_Lat,
        'Long': list1_Long,
        'UTM_X': list1_UTM_X,
        'UTM_Y': list1_UTM_Y,
        'Species': list1_species,
        'Quantity': list1_quantity
    }
)
df_Field1_Converted

# Save this dataframe into local csv file! 
df_Field1_Converted[['Plot','Lat','Long','Species','Quantity']].to_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid LatLong with Species.csv", index = False)
df_Field1_Converted[['Plot','UTM_X','UTM_Y','Species','Quantity']].to_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid UTM with Species.csv", index = False)
df_Field1_Converted[['Plot','Species','Quantity']].to_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid with Species.csv", index = False)

### Retrieve Typology

# Read typology first
df_Field1_Typo = df_Field1.copy()
df_Field1_Typo = df_Field1_Typo[['ESU','Tipologia Forestale castrovinci']]
# Keep only valid plots
df_Field1_Typo = pd.merge(df_Field1_Typo,df_Field1_Merge['ESU'],on='ESU',how='inner')
# Rename Columns
df_Field1_Typo.columns = ['Plot','Typology']
df_Field1_Typo

# Save!
df_Field1_Typo.to_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid Typology.csv", index = False)

### Merge Association

df_Field1_Asso = pd.read_csv(cwd_Field_Data1 + "\\ROI_30x30_rilievo_floristico_syntax_&_associazioni_forestali.csv")
df_Field1_Asso.tail()

df_Field1_Asso = df_Field1_Asso.sort_values('ESU')
df_Field1_Asso

df_Field1_Asso = df_Field1_Asso.rename(columns = {'ESU': 'Plot', 'Associazio': "Association"})
df_Field1_Asso[['Plot','Association']].to_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid Association.csv", index = False)
