==> Python/5-2-FieldData2 to Shapefile.ipynb <==
# This notebook is aimed at the creation of shapefiles of field data 1, based on the .csv files which we already created in the last step. 

import pandas as pd
import geopandas as gpd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Field_Data1 = cwd_Field + '\\Field Dataset 1'
cwd_Field_Data1_CSV = cwd_Field_Data1 + "\\CSV"
cwd_Field_Data1_Shapefiles = cwd_Field_Data1 + "\\Shapefiles"

### Read our perfect .csv files

df_field_Valid_LatLong = pd.read_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid LatLong.csv")
df_field_Valid_LatLong.head()

df_field_Valid_UTM = pd.read_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid UTM.csv")
df_field_Valid_UTM.head()

### Create geodataframes

gdf_field_Valid_LatLong = gpd.GeoDataFrame(
    df_field_Valid_LatLong,
    geometry = gpd.points_from_xy(df_field_Valid_LatLong['Long'],df_field_Valid_LatLong['Lat']),
    crs = 'EPSG:4326'
)
gdf_field_Valid_LatLong

gdf_field_Valid_UTM = gpd.GeoDataFrame(
    df_field_Valid_UTM,
    geometry = gpd.points_from_xy(df_field_Valid_UTM['UTM_X'],df_field_Valid_UTM['UTM_Y']),
    crs = 'EPSG:32632'
)
gdf_field_Valid_UTM

# Save them as shapefiles! 
gdf_field_Valid_LatLong.to_file(cwd_Field_Data1_Shapefiles + "\\FieldData1 Valid LatLong.shp")
gdf_field_Valid_UTM.to_file(cwd_Field_Data1_Shapefiles + "\\FieldData1 Valid UTM.shp")
