==> Python/6-FieldData2 to Shapefile.ipynb <==
import numpy as np
import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Field_Data2 = cwd_Field + '\\Field Dataset 2'

# Read the Excel file which contains all the geographical coordinates of the field plots! 
df_Field_Coord = pd.read_excel(cwd_Field_Data2 + "\\Dati coordinate plot forestali Parco Ticino.xlsx")
df_Field_Coord

# Immediately create our shapefile from this excel sheet
gdf_Field_All_4326 = gpd.GeoDataFrame(
    df_Field_Coord['n° plot'],
    geometry = gpd.points_from_xy(df_Field_Coord['X'],df_Field_Coord['Y']),
    crs = 'EPSG:4326'
)
# Create another copy but with UTM crs
gdf_Field_All_32632 = gdf_Field_All_4326.to_crs(32632)
# Add the geographical coordinates to the columns and save to local shapefiles! 
gdf_Field_All_4326['Lat'] = df_Field_Coord['Y']
gdf_Field_All_4326['Long'] = df_Field_Coord['X']
gdf_Field_All_32632['UTM_X'] = gdf_Field_All_32632.geometry.x
gdf_Field_All_32632['UTM_Y'] = gdf_Field_All_32632.geometry.y
gdf_Field_All_4326

# Save the these geodataframes into local shapefiles! 
gdf_Field_All_4326.to_file(cwd_Field + "\\Field Data 2 All Points (4326).shp")
gdf_Field_All_32632.to_file(cwd_Field + "\\Field Data 2 All Points (32632).shp")

# However there are some plots outside our RoI. Therefore we should remove those invalid plots. 
# Load our region of interest shapefile! 
gdf_RoI = gpd.read_file(cwd_Field + "\\ShapefileCorretti\\confini_foreste_corretti.shp")

# Check which field plots (points) are within our RoI! And save their plot numbers into a list
gdf_InnerJoin = gdf_RoI.sjoin(gdf_Field_All_32632, how="inner").sort_values('n° plot')
list_ValidPlots = pd.unique(gdf_InnerJoin['n° plot'])
list_ValidPlots

# Only keep the valid plots and recreate another new dataframe which only contains the field plots inside our RoI! 
df_Field_Valid = df_Field_Coord.copy()
df_Field_Valid = df_Field_Valid.where(df_Field_Valid['n° plot'].isin(list_ValidPlots)).dropna().reset_index(drop = True)
df_Field_Valid['n° plot'] = df_Field_Valid['n° plot'].astype('Int64')
df_Field_Valid

# Now we create new geodataframes from our new dataframes
gdf_Field_Valid_4326 = gpd.GeoDataFrame(
    df_Field_Valid['n° plot'],
    geometry = gpd.points_from_xy(df_Field_Valid['X'],df_Field_Valid['Y']),
    crs = 'EPSG:4326'
)
# Convert the crs from 4326 (WGS84) to 32632 (WGS84 UTM zone 32N)
gdf_Field_Valid_32632 = gdf_Field_Valid_4326.to_crs(32632)
# Add the geographical coordinates to the columns and save to local shapefiles! 
gdf_Field_Valid_4326['Lat'] = df_Field_Valid['Y']
gdf_Field_Valid_4326['Long'] = df_Field_Valid['X']
gdf_Field_Valid_4326.to_file(cwd_Field + "\\Field Data 2 Valid Points (4326).shp")
gdf_Field_Valid_32632['UTM_X'] = gdf_Field_Valid_32632.geometry.x
gdf_Field_Valid_32632['UTM_Y'] = gdf_Field_Valid_32632.geometry.y
gdf_Field_Valid_32632.to_file(cwd_Field + "\\Field Data 2 Valid Points (32632).shp")

# OnHold! Calculation of Index!

# Since our Excel file contains multiple sheets, we need to get the names of sheets firstly. 
xlsx = pd.ExcelFile(cwd_Field_Data2 + '\\Dati plot forestali Parco Ticino completo.xlsx')
xlsx.sheet_names

# Combine! 
xlsx_field = pd.DataFrame()
for sheet in xlsx.sheet_names:
    if sheet != 'TEMPLATE':
        temp_df = pd.read_excel(cwd_Field_Data2 + '\\Dati plot forestali Parco Ticino completo.xlsx', sheet_name=sheet)
        xlsx_field = pd.concat([xlsx_field, temp_df], ignore_index=True)
xlsx_field

# By observing our concatenated dataframe, we only need to keep the first 3 columns. 
# Drop those columns! 
xlsx_field = xlsx_field.drop(xlsx_field.iloc[:,np.r_[1,np.arange(3,6,1)]], axis = 1)
xlsx_field

# Now we need to drop the useless rows. If the first column of a row is not a numeric value, then it will be dropped. 
df_field = xlsx_field.copy()
for row in range(0,xlsx_field.shape[0]):
    # Get the index label of the current row! 
    temp_label = xlsx_field.iloc[row].name
    if str(xlsx_field.iloc[row,0]).isnumeric(): # Note: we have to convert the value to str first, otherwise we can't apply "isnumeric()" function. 
        continue
    else: 
        # Drop the row based on the label! 
        df_field = df_field.drop(index=temp_label, axis = 0)
# Reset index
df_field = df_field.reset_index(drop = True)
# Assign column names
df_field.columns = ["Plot Number", "Tree Species"]
# Finally we get our clean dataframe! 
df_field
