==> Python/6-1-FieldData2 to Shapefile and filter.ipynb <==
# This notebook is dedicated to convert the field data from another group into shapefiles. Besides, there is also a step to filter out the plots outside our RoI. 
For the field data 2, since there are several plots outside our RoI, we need to filter our them first, hence the conversion to .csv step being doable after this step

import pandas as pd
import geopandas as gpd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Field_Data2 = cwd_Field + '\\Field Dataset 2'
cwd_Field_Data2_CSV = cwd_Field_Data2 + "\\CSV"
cwd_Field_Data2_Shapefiles = cwd_Field_Data2 + "\\Shapefiles"

### Refine a bit the Excel file of the geographical coordinates of all the field plots and save it as .csv file

# Read the Excel file which contains all the geographical coordinates of the field plots! 
df_Field_Coord = pd.read_excel(cwd_Field_Data2 + "\\Dati coordinate plot forestali Parco Ticino.xlsx")
df_Field_Coord.head()

# Refine a bit this dataframe and then save it as .csv file.
df_Field_Coord.columns = ['Plot','Lat','Long']
df_Field_Coord.to_csv(cwd_Field_Data2_CSV + '\\FieldData2 All LatLong.csv', index = False)

### Create a geodataframe and save it as shapefile

df_Field = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 All LatLong.csv")
# Immediately create our shapefile from this excel sheet
gdf_Field_All_4326 = gpd.GeoDataFrame(
    df_Field,
    geometry = gpd.points_from_xy(df_Field['Long'],df_Field['Lat']),
    crs = 'EPSG:4326'
)
gdf_Field_All_4326.head()

gdf_Field_All_4326.to_file(cwd_Field_Data2_Shapefiles + "\\FieldData2 All LatLong.shp")

### Also create a shapefile of UTM crs (32632). 

gdf_Field_All_32632 = gdf_Field_All_4326[['Plot','geometry']].to_crs(32632)
gdf_Field_All_32632['UTM_X'] = gdf_Field_All_32632.geometry.x
gdf_Field_All_32632['UTM_Y'] = gdf_Field_All_32632.geometry.y
gdf_Field_All_32632 = gdf_Field_All_32632[['Plot','UTM_X','UTM_Y','geometry']]
gdf_Field_All_32632.head()

gdf_Field_All_32632.to_file(cwd_Field_Data2_Shapefiles + "\\FieldData2 All UTM.shp")

### Meanwhile, also save this geodataframe of UTM crs as .csv file

df_Field_UTM = gdf_Field_All_32632[['Plot','UTM_X','UTM_Y']]
df_Field_UTM.to_csv(cwd_Field_Data2_CSV + '\\FieldData2 All UTM.csv', index = False)

### Now we need to filter out the plots outside our RoI. 

# Load our region of interest shapefile! 
gdf_RoI = gpd.read_file(cwd_Field + "\\ShapefileCorretti\\confini_foreste_corretti.shp")

# Check which field plots (points) are within our RoI! And save their plot numbers into an array
gdf_InnerJoin = gdf_RoI.sjoin(gdf_Field_All_32632, how="inner").sort_values('Plot')
arr_ValidPlots = pd.unique(gdf_InnerJoin['Plot'])
arr_ValidPlots

# Only keep the valid plots and recreate another new dataframe which only contains the field plots inside our RoI! 
df_Field_Valid = df_Field_Coord.copy()
mask_valid = df_Field_Valid['Plot'].isin(arr_ValidPlots)
df_Field_Valid = df_Field_Valid[mask_valid]
df_Field_Valid = df_Field_Valid.reset_index(drop = True)
df_Field_Valid.tail()

### Now save this dataframe as .csv file

df_Field_Valid_4326 = df_Field_Valid.copy()
df_Field_Valid_4326.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid LatLong.csv", index = False)

### Then, recreate a geodataframe from this filtered dataframe and save it as shapefile. 

gdf_Field_Valid_4326 = gpd.GeoDataFrame(
    df_Field_Valid_4326,
    geometry = gpd.points_from_xy(df_Field_Valid_4326['Long'],df_Field_Valid_4326['Lat']),
    crs = 'EPSG:4326'
)
gdf_Field_Valid_4326.tail()

gdf_Field_Valid_4326.to_file(cwd_Field_Data2_Shapefiles + "\\FieldData2 Valid LatLong.shp")

### And again, convert the crs to UTM and save it as shapefile as well as the csv file. 

gdf_Field_Valid_32632 = gdf_Field_Valid_4326[['Plot','geometry']].copy().to_crs(32632)
gdf_Field_Valid_32632['UTM_X'] = gdf_Field_Valid_32632.geometry.x
gdf_Field_Valid_32632['UTM_Y'] = gdf_Field_Valid_32632.geometry.y
gdf_Field_Valid_32632 = gdf_Field_Valid_32632[['Plot','UTM_X','UTM_Y','geometry']]
gdf_Field_Valid_32632.tail()

gdf_Field_Valid_32632.to_file(cwd_Field_Data2_Shapefiles + "\\FieldData2 Valid UTM.shp")

df_Field_Valid_32632 = gdf_Field_Valid_32632[['Plot','UTM_X','UTM_Y']]
df_Field_Valid_32632.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid UTM.csv", index = False)
