==> Python/6-2-FieldData2 to CSV.ipynb <==
# This notebook illustrates the workflow for converting field data from another group in .xlsx file format to .csv files, keeping only the necessary data.

import numpy as np
import pandas as pd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Field_Data2 = cwd_Field + '\\Field Dataset 2'
cwd_Field_Data2_CSV = cwd_Field_Data2 + "\\CSV"

### Purge the raw Excel file

# Since our Excel file contains multiple sheets, we need to get the names of sheets firstly. 
xlsx = pd.ExcelFile(cwd_Field_Data2 + '\\Dati plot forestali Parco Ticino completo.xlsx')
xlsx.sheet_names

# Combine these sheets into a single one! 
xlsx_Field = pd.DataFrame()
for sheet in xlsx.sheet_names:
    if sheet != 'TEMPLATE':
        temp_df = pd.read_excel(cwd_Field_Data2 + '\\Dati plot forestali Parco Ticino completo.xlsx', sheet_name=sheet)
        xlsx_Field = pd.concat([xlsx_Field, temp_df], ignore_index=True)
xlsx_Field

# By observing our concatenated dataframe, we only need to keep the first 3 columns. 
# Drop those columns! 
xlsx_Field = xlsx_Field.drop(xlsx_Field.iloc[:,np.r_[1,np.arange(3,6,1)]], axis = 1)
xlsx_Field

# Now we need to drop the useless rows. If the first column of a row is not a numeric value, then it will be dropped. 
df_Field = xlsx_Field.copy()
for row in range(0,xlsx_Field.shape[0]):
    # Get the index label of the current row! 
    temp_label = xlsx_Field.iloc[row].name
    if str(xlsx_Field.iloc[row,0]).isnumeric(): # Note: we have to convert the value to str first, otherwise we can't apply "isnumeric()" function. 
        continue
    else: 
        # Drop the row based on the label! 
        df_Field = df_Field.drop(index=temp_label, axis = 0)
# Reset index
df_Field = df_Field.reset_index(drop = True)
# Assign column names
df_Field.columns = ["Plot", "Species"]
# Finally we get our clean dataframe! 
df_Field

# Now we group by to aggregate some values
df_Field_Count = df_Field.groupby(['Plot', 'Species'], as_index = False).size()
df_Field_Count = df_Field_Count.rename(columns = {'size': 'Quantity'})
df_Field_Count

### Save this refined dataframe as .csv file. 

df_Field_Count.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 All with Species.csv", index = False)

### Merge the geographical coordinates of the plots and also save them as .csv files

# Read another Excel file which contains all the geographical coordinates of the field plots! 
df_Field_All_LatLong = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 All LatLong.csv")
df_Field_All_UTM = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 All UTM.csv")

df_FieldMerge_All_LatLong = pd.merge(df_Field_Count,df_Field_All_LatLong,on = 'Plot')
df_FieldMerge_All_UTM = pd.merge(df_Field_Count,df_Field_All_UTM,on = 'Plot')
df_FieldMerge_All_UTM

df_FieldMerge_All_LatLong.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 All LatLong with Species.csv")
df_FieldMerge_All_UTM.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 All UTM with Species.csv")

### Now we filter out the plots outside our RoI

# Read the filtered out .csv file
df_Field_Valid = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid LatLong.csv")
df_Field_Valid.tail()

# Get an array of the numbers of the plots inside our RoI
array_Valid = df_Field_Valid['Plot'].unique()
array_Valid

# Mask the invalid plots
mask_Valid = df_Field_Count['Plot'].isin(array_Valid)
df_Field_Count_Valid = df_Field_Count[mask_Valid].reset_index(drop = True)
df_Field_Count_Valid

# Mask the invalid plots
mask_Valid = df_FieldMerge_All_LatLong['Plot'].isin(array_Valid)
df_Field_Valid_LatLong = df_FieldMerge_All_LatLong[mask_Valid].reset_index(drop = True)
df_Field_Valid_LatLong

# Mask the invalid plots
mask_Valid = df_FieldMerge_All_UTM['Plot'].isin(array_Valid)
df_Field_Valid_UTM = df_FieldMerge_All_UTM[mask_Valid].reset_index(drop = True)
df_Field_Valid_UTM

### Save these filtered out dataframes as .csv files

df_Field_Count_Valid.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid with Species.csv", index = False)
df_Field_Valid_UTM.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid UTM with Species.csv", index = False)
df_Field_Valid_LatLong.to_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid LatLong with Species.csv", index = False)
