==> Python/7-1-FieldData1 2 Merge and Export.ipynb <==
# This notebook simply merges the dataset 1 and dataset 2 into a single one and export as .csv and shapefiles. 

import numpy as np
import pandas as pd
import geopandas as gpd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Field_Data1 = cwd_Field + '\\Field Dataset 1'
cwd_Field_Data1_CSV = cwd_Field_Data1 + "\\CSV"
cwd_Field_Data1_Shapefiles = cwd_Field_Data1 + "\\Shapefiles"
cwd_Field_Data2 = cwd_Field + '\\Field Dataset 2'
cwd_Field_Data2_CSV = cwd_Field_Data2 + "\\CSV"
cwd_Field_Data2_Shapefiles = cwd_Field_Data2 + "\\Shapefiles"
cwd_Field_DataMerged = cwd_Field + '\\Field Dataset Merged'
cwd_Field_DataMerged_CSV = cwd_Field_DataMerged + "\\CSV"
cwd_Field_DataMerged_Shapefiles = cwd_Field_DataMerged + "\\Shapefiles"

### First import the csv files created in step 5 & 6 of each field dataset as dataframes

df_Field1_LatLong = pd.read_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid LatLong.csv")
df_Field1_UTM = pd.read_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid UTM.csv")
df_Field1_UTM.head()

df_Field2_LatLong = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid LatLong.csv")
df_Field2_UTM = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid UTM.csv")
df_Field2_UTM.head()

### Simply concatenate these two dataframes

df_Field_LatLong = pd.concat([df_Field1_LatLong,df_Field2_LatLong])
df_Field_LatLong

df_Field_UTM = pd.concat([df_Field1_UTM,df_Field2_UTM])
df_Field_UTM

# Save as .csv files! 
df_Field_LatLong.to_csv(cwd_Field_DataMerged_CSV + "\\FieldDataMerged Valid LatLong.csv", index = False)
df_Field_UTM.to_csv(cwd_Field_DataMerged_CSV + "\\FieldDataMerged Valid UTM.csv", index = False)

### Now we transfer these "merged" dataframes into geodataframes

gdf_Field_LatLong = gpd.GeoDataFrame(
    df_Field_LatLong,
    geometry = gpd.points_from_xy(df_Field_LatLong['Long'],df_Field_LatLong['Lat']),
    crs = 'EPSG:4326'
)
gdf_Field_LatLong

gdf_Field_UTM = gpd.GeoDataFrame(
    df_Field_UTM,
    geometry = gpd.points_from_xy(df_Field_UTM['UTM_X'],df_Field_UTM['UTM_Y']),
    crs = 'EPSG:32632'
)
gdf_Field_UTM

### Before further process, we save these merged geodataframes as shapefiles in case that they are needed in the future.

gdf_Field_LatLong.to_file(cwd_Field_DataMerged_Shapefiles + "\\FieldDataMerged Valid LatLong.shp")
gdf_Field_UTM.to_file(cwd_Field_DataMerged_Shapefiles + "\\FieldDataMerged Valid UTM.shp")

### Now we combine the .csv files with species data

df_Field1_Species_LatLong = pd.read_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid LatLong with Species.csv")
df_Field1_Species_UTM = pd.read_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid UTM with Species.csv")
df_Field2_Species_LatLong = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid LatLong with Species.csv")
df_Field2_Species_UTM = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid UTM with Species.csv")
df_Field1_Species = pd.read_csv(cwd_Field_Data1_CSV + "\\FieldData1 Valid with Species.csv")
df_Field2_Species = pd.read_csv(cwd_Field_Data2_CSV + "\\FieldData2 Valid with Species.csv")

df_Field_Species_LatLong = pd.concat([df_Field1_Species_LatLong,df_Field2_Species_LatLong])
df_Field_Species_LatLong

df_Field_Species_UTM = pd.concat([df_Field1_Species_UTM,df_Field2_Species_UTM])
df_Field_Species_UTM

df_Field_Species = pd.concat([df_Field1_Species,df_Field2_Species])
df_Field_Species

df_Field_Species_LatLong.to_csv(cwd_Field_DataMerged_CSV + "\\FieldDataMerged Valid LatLong with Species.csv", index = False)
df_Field_Species_UTM.to_csv(cwd_Field_DataMerged_CSV + "\\FieldDataMerged Valid UTM with Species.csv", index = False)
df_Field_Species.to_csv(cwd_Field_DataMerged_CSV + "\\FieldDataMerged Valid with Species.csv", index = False)
