==> Python/7-2-FieldDataMerged Buffer.ipynb <==
# This notebook simply adds some buffer (5 x 10m for sentinel-2 and 5 x 30m for PRISMA) to our field data, in order for the R package biodivMapR

import geopandas as gpd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Field_DataMerged = cwd_Field + '\\Field Dataset Merged'
cwd_Field_DataMerged_Shapefiles = cwd_Field_DataMerged + "\\Shapefiles"
cwd_Field_DataMerged_Shapefiles_Buffer = cwd_Field_DataMerged + "\\Shapefiles Buffered"

gdf_Field_Valid_UTM = gpd.read_file(cwd_Field_DataMerged_Shapefiles + "\\FieldDataMerged Valid UTM.shp")
gdf_Field_Valid_UTM

# Add 5 x 10m square buffer for Sentinel 2. 
gdf_Field_Valid_UTM_Buffer50 = gdf_Field_Valid_UTM.copy()
gdf_Field_Valid_UTM_Buffer50['geometry'] = gdf_Field_Valid_UTM_Buffer50.geometry.buffer(25, cap_style = 3) # cap_style = 1 by default but for circular buffer
gdf_Field_Valid_UTM_Buffer50

# Add 5 x 30m square buffer for PRISMA. 
gdf_Field_Valid_UTM_Buffer150 = gdf_Field_Valid_UTM.copy()
gdf_Field_Valid_UTM_Buffer150['geometry'] = gdf_Field_Valid_UTM_Buffer150.geometry.buffer(75, cap_style = 3)
gdf_Field_Valid_UTM_Buffer150

# Add 10 x 10m square buffer for Sentinel 2. 
gdf_Field_Valid_UTM_Buffer100 = gdf_Field_Valid_UTM.copy()
gdf_Field_Valid_UTM_Buffer100['geometry'] = gdf_Field_Valid_UTM_Buffer100.geometry.buffer(50, cap_style = 3)
gdf_Field_Valid_UTM_Buffer100

# Add 10 x 30m square buffer for PRISMA. 
gdf_Field_Valid_UTM_Buffer300 = gdf_Field_Valid_UTM.copy()
gdf_Field_Valid_UTM_Buffer300['geometry'] = gdf_Field_Valid_UTM_Buffer300.geometry.buffer(150, cap_style = 3)
gdf_Field_Valid_UTM_Buffer300

# Add 10 x 30m square buffer for PRISMA. 
gdf_Field_Valid_UTM_Buffer30 = gdf_Field_Valid_UTM.copy()
gdf_Field_Valid_UTM_Buffer30['geometry'] = gdf_Field_Valid_UTM_Buffer30.geometry.buffer(15, cap_style = 3)
gdf_Field_Valid_UTM_Buffer30

gdf_Field_Valid_UTM_Buffer150.iloc[0:63,:]

# Save these buffered geodataframes as shapefiles! 

gdf_Field_Valid_UTM_Buffer50.to_file(cwd_Field_DataMerged_Shapefiles_Buffer + "\\FieldDataMerged Valid Buffer 50m UTM.shp")
gdf_Field_Valid_UTM_Buffer150.to_file(cwd_Field_DataMerged_Shapefiles_Buffer + "\\FieldDataMerged Valid Buffer 150m UTM.shp")
gdf_Field_Valid_UTM_Buffer100.to_file(cwd_Field_DataMerged_Shapefiles_Buffer + "\\FieldDataMerged Valid Buffer 100m UTM.shp")

gdf_Field_Valid_UTM_Buffer300.to_file(cwd_Field_DataMerged_Shapefiles_Buffer + "\\FieldDataMerged Valid Buffer 300m UTM.shp")
