==> Python/9-Raster Indices Post Process.ipynb <==
# This notebook is dedicated at post processing those .csv files generated in R Studio. 

To be more specific, this notebook extracts the useful data from those .csv files and merge them into a single .csv file, for easier future use. 

import numpy as np
import pandas as pd

# Set some directories. Here we use absolute directories. 
cwd = 'c:\\Users\\m1865\\Desktop\\Ticino'
cwd_Field = cwd + '\\FieldData'
cwd_Raster = cwd + "\\Results"

# Write here the clusters we used in R Studio
list_Cluster_Number = [20,35,50]
# Write here the names of the raster
list_Name_Raster = ['2022 06Cropped','2022 09Cropped']

# Get the folder names separated by clusters
list_Cluster = []
for item in list_Cluster_Number:
    temp = str(item) + "Clusters"
    list_Cluster.append(temp)
list_Cluster

### Post process of Alpha Indices Window Size 5 (Default)

# Save only useful columns and save them to local storage as normal .csv files! 
for i in list_Name_Raster:
    df_Raster_Month = pd.DataFrame()
    for j in list_Cluster:
        cwd_temp = cwd_Raster + "\\" + j + "\\" + i + "\\AlphaDiversity.csv"
        df_temp = pd.read_csv(cwd_temp, sep = '\t')
        df_temp = df_temp[['Plot','Species_Richness','Shannon','Simpson']]
        df_temp.rename(columns = {
            'Species_Richness': "Raster " + j + " Species Richness",
            "Shannon": "Raster " + j + " Shannon",
            "Simpson": "Raster " + j + " Simpson"
        }, inplace = True)
        if df_Raster_Month.empty:
            df_Raster_Month = df_temp.copy()
        else:
            df_Raster_Month = pd.merge(df_Raster_Month, df_temp, on = 'Plot')
    df_Raster_Month.to_csv(cwd_Raster + "\\" + i + " Alpha Diversity.csv", index = False)

### Post process of Beta Bray-Curtis Index Window Size 5 (Default)

# Only get the upper triangle elements without diagonal ones. And then convert them into an one-dimensional array and merge them. 
for i in list_Name_Raster:
    df_Raster_Month = pd.DataFrame()
    for j in list_Cluster:
        cwd_temp = cwd_Raster + "\\" + j + "\\" + i + "\\BrayCurtis.csv"
        df_temp = pd.read_csv(cwd_temp, sep = '\t')
        arr_temp = df_temp.to_numpy()
        arr_temp = arr_temp[np.triu_indices_from(arr_temp, k = 1)]
        if df_Raster_Month.empty:
            df_Raster_Month = pd.DataFrame(
                arr_temp,
                columns = ["Raster " + j + " Beta"]
            )
        else:
            df_Raster_Month["Raster " + j + " Beta"] = arr_temp
    df_Raster_Month.to_csv(cwd_Raster + "\\" + i + " Beta Diversity.csv", index = False)

## Window Size 10 (Recommended)
### Alpha

# Save only useful columns and save them to local storage as normal .csv files! 
for i in list_Name_Raster:
    df_Raster_Month = pd.DataFrame()
    for j in list_Cluster:
        cwd_temp = cwd_Raster + "\\" + j + "\\" + i + "\\AlphaDiversity10.csv"
        df_temp = pd.read_csv(cwd_temp, sep = '\t')
        df_temp = df_temp[['Plot','Species_Richness','Shannon','Simpson']]
        df_temp.rename(columns = {
            'Species_Richness': "Raster " + j + " Species Richness",
            "Shannon": "Raster " + j + " Shannon",
            "Simpson": "Raster " + j + " Simpson"
        }, inplace = True)
        if df_Raster_Month.empty:
            df_Raster_Month = df_temp.copy()
        else:
            df_Raster_Month = pd.merge(df_Raster_Month, df_temp, on = 'Plot')
    df_Raster_Month.to_csv(cwd_Raster + "\\" + i + " Alpha Diversity10.csv", index = False)

### Beta

# Only get the upper triangle elements without diagonal ones. And then convert them into an one-dimensional array and merge them. 
for i in list_Name_Raster:
    df_Raster_Month = pd.DataFrame()
    for j in list_Cluster:
        cwd_temp = cwd_Raster + "\\" + j + "\\" + i + "\\BrayCurtis10.csv"
        df_temp = pd.read_csv(cwd_temp, sep = '\t')
        arr_temp = df_temp.to_numpy()
        arr_temp = arr_temp[np.triu_indices_from(arr_temp, k = 1)]
        if df_Raster_Month.empty:
            df_Raster_Month = pd.DataFrame(
                arr_temp,
                columns = ["Raster " + j + " Beta"]
            )
        else:
            df_Raster_Month["Raster " + j + " Beta"] = arr_temp
    df_Raster_Month.to_csv(cwd_Raster + "\\" + i + " Beta Diversity10.csv", index = False)
